Move herd-specific files into new package and rename them
